Test/DBTests/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
strawberry-graphql[fastapi]==0.254.0
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.7.0
